siguen los arreglos

+++ src/main/java/Vista/TransportGUI.java
@@ -16,7 +16,7 @@ import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
 import java.util.ArrayList;
 
-public class TransportGUI {
+public class TransportGUI extends JPanel{
 
 
     private JPanel panel1;
@@ -39,7 +39,7 @@ public class TransportGUI {
 
     public TransportGUI() {
         this.initialize();
-        this.tableInit();
+        //this.tableInit();
     }
 
 
@@ -74,6 +74,7 @@ public class TransportGUI {
                 tadd.setAnterior(TransportGUI.this.frameTransport);
                 tadd.frameTransportAdd.setVisible(true);
 
+                frameTransport.dispose();
 
 
             }
@@ -111,8 +112,11 @@ public class TransportGUI {
                     JOptionPane.showMessageDialog(null, "Seleccione una estacion para eliminar.",
                             "ADVERTENCIA", JOptionPane.ERROR_MESSAGE);
                 } else {
-                    deleteTransport();
-                    JOptionPane.showMessageDialog(frameTransport,"Estacion eliminada" , "Failure", JOptionPane.INFORMATION_MESSAGE);
+                    int resp=  JOptionPane.showConfirmDialog(null, "¿Esta seguro de eliminar? Considere que se quitaran las rutas a dicho transporte");
+                    if(JOptionPane.OK_OPTION == resp) {
+                        deleteTransport();
+                    }
+
                 }
 
             }
@@ -126,23 +130,32 @@ public class TransportGUI {
 
         String[] atributosTransporte = {"Id", "Nombre", "Color", "Estado"};
         CBsearchTransport.setModel(new DefaultComboBoxModel<String>(atributosTransporte));
-        DTOTransport transportParam = new DTOTransport();
+
 
         buscarButton.addActionListener(new ActionListener() {
             public void actionPerformed(ActionEvent e) {
                 String param = textTransport.getText();
+                DTOTransport transportParam = new DTOTransport();
+
 
                 switch (CBsearchTransport.getSelectedIndex()){
                     case 0: {
                         //busqueda por id
                         Integer id = Integer.parseInt(param);
+                        try{
+                            id = Integer.parseInt(param);
+                        }catch(NumberFormatException i){
+                            JOptionPane.showMessageDialog(null, "El campo tiene que ser numerico",
+                                    "ADVERTENCIA", JOptionPane.ERROR_MESSAGE);
+                            return;
+                        }
                         transportParam.setIdTransport(id);
 
                         break;
                     }
                     case 1:{
                         //busca por nombre
-                        transportParam.setName(param);
+                        transportParam.setName(param.substring(0,1).toUpperCase() + param.substring(1).toLowerCase());
 
                         break;
                     }
@@ -167,7 +180,14 @@ public class TransportGUI {
                     default:
                 }
                 ArrayList<DTOTransport> result = tm.searchDTOTransport(transportParam);
-                updateTabla(result);
+                if(result.size()>0){
+                    updateTabla(result);
+                }else{
+                    JOptionPane.showMessageDialog(null, "No se encuentran transportes con los atributos seleccionados.",
+                            "ADVERTENCIA", JOptionPane.ERROR_MESSAGE);
+
+                }
+
 
             }
         });
@@ -185,7 +205,7 @@ public class TransportGUI {
                     CBStatus.setVisible(false);
                     textTransport.setVisible(true);
                 }
-
+                table.repaint();
             }
         });
 
@@ -198,7 +218,10 @@ public class TransportGUI {
         int id = Integer.parseInt(this.table.getModel().getValueAt(selected, 0).toString());
         DTOTransport deleteT = new DTOTransport();
         deleteT.setIdTransport(id);
-        TransportManager.deleteTransportRoute(deleteT);
+        tm.deleteTransportRoute(deleteT);
+        JOptionPane.showMessageDialog(frameTransport,"Estacion eliminada" , "Failure", JOptionPane.INFORMATION_MESSAGE);
+        TransportGUI.this.anterior.setVisible(true);
+        TransportGUI.this.frameTransport.dispose();
         table.repaint();
 
     }
@@ -223,8 +246,9 @@ public class TransportGUI {
         }
         table.setModel(t);
 
+
     }
-    public void tableInit(){
+  /*  public void tableInit(){
         String [] atributos = {"ID", "Nombre", "Color", "Estado"};
         DefaultTableModel model = (DefaultTableModel)this.table.getModel();
         model.setRowCount(0);
@@ -241,10 +265,10 @@ public class TransportGUI {
             }
             Object[] datos = {id, name, color, estado};
             model.addRow(datos);
-
+            table.repaint();
         }
     }
-
+*/
